+++ src/components/policy/policyApis/policyApis.ts
@@ -203,7 +203,7 @@ export const createPoliciesWithBenefits = (
         BTerm: parseInt(benefits?.BTerm),
         BPTerm: parseInt(benefits?.BPTerm),
         BSumAssured: parseInt(benefits?.BSumAssured),
-        Interest: parseInt(benefits?.Interest),
+        Interest: parseFloat(benefits?.Interest),
       })),
     },
